requirements.txt
discord.py==1.7.3
torch==1.10.1
setuptools==59.5.0
transformers==4.16.2
python-dotenv==0.20.0
